+++ packages/compiler/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.13.1](https://github.com/vibeunion/supacloud/compare/compiler-v0.13.0...compiler-v0.13.1) (2026-09-08)
+
+
+### Bug Fixes
+
+* **compiler:** avoid duplicate GraphQL enum and input types ([#1228](https://github.com/vibeunion/supacloud/issues/1228)) ([9054c41](https://github.com/vibeunion/supacloud/commit/9054c417b3e16c2d265ea2839f4aa72f9a64f513))
+
+## [0.13.0](https://github.com/vibeunion/supacloud/compare/compiler-v0.12.0...compiler-v0.13.0) (2026-09-08)
+
+
+### Features
+
+* **compiler:** add database-first GraphQL query contracts ([#1224](https://github.com/vibeunion/supacloud/issues/1224)) ([a10e0d5](https://github.com/vibeunion/supacloud/commit/a10e0d54f87bee79c0a3a34f9b8685a994eb444b))
+
 ## [0.12.0](https://github.com/vibeunion/supacloud/compare/compiler-v0.11.0...compiler-v0.12.0) (2026-09-08)